prediction logic depended on dt
